Update home page carousels
Create a reusable Carousel component that can be shared
